Move feature types into feature/types.ts file
