+++ frontend/src/common/layout/AppHeader.tsx
@@ -9,7 +9,7 @@ import { Header3 } from "@/components/ui/typography";
 import { useAuth } from "@/features/auth/hooks/useAuth";
 import { ManageUserSheet } from "@/features/user/components/ManageUserSheet";
 import { useProfile } from "@/features/user/hooks/useProfile";
-import { UserGroup } from "@/types/user";
+import { UserGroup } from "@/features/user/types";
 import { BellIcon, MailsIcon, PlusSquareIcon } from "lucide-react";
 
 const AdminNavigation = () => (
