+++ frontend/src/features/approvals/datatable/ApprovalListColumns.tsx
@@ -1,4 +1,4 @@
-import { ApprovalResponse } from "@/types/task";
+import { ApprovalResponse } from "@/features/tasks/types";
 import { ColumnDef } from "@tanstack/react-table";
 
 export const approvalListColumns: ColumnDef<ApprovalResponse>[] = [
