+++ frontend/src/features/auth/AuthContext.tsx
@@ -1,4 +1,4 @@
-import { User } from "@/types/user";
+import { User } from "@/features/user/types";
 import React from "react";
 
 export interface AuthContext {
