+++ frontend/src/features/auth/components/LoginForm.tsx
@@ -18,8 +18,8 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { useAuth } from "@/features/auth/hooks/useAuth";
+import { User } from "@/features/user/types";
 import { Route } from "@/routes/(auth)/login";
-import { User } from "@/types/user";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useNavigate, useRouter } from "@tanstack/react-router";
 import { useState } from "react";
