+++ frontend/src/features/kids/components/MyActiveTasksCard.tsx
@@ -2,7 +2,7 @@ import { Card } from "@/components/ui/card";
 import { Spinner } from "@/components/ui/spinner";
 import { StatLabel } from "@/components/ui/typography";
 import { AssignedTaskCard } from "@/features/kids/components/AssignedTaskCard";
-import { AssignedTask } from "@/types/task";
+import { AssignedTask } from "@/features/tasks/types";
 
 interface MyActiveTasksCardProps {
   assignedTasks: AssignedTask[] | undefined;
