+++ frontend/src/features/parents/components/TaskApprovalCard.tsx
@@ -1,7 +1,7 @@
 import { CardContent, CardDescription, CardTitle } from "@/components/ui/card";
 import { ApprovalRowActions } from "@/features/parents/components/ApprovalRowActions";
 import TaskCategoryBadge from "@/features/tasks/components/TaskCategoryBadge";
-import { ApprovalResponse } from "@/types/task";
+import { ApprovalResponse } from "@/features/tasks/types";
 
 export interface TaskApprovalCardProps {
   approval: ApprovalResponse;
