+++ frontend/src/features/rewards/types.ts
