+++ frontend/src/features/tasks/components/TaskApprovalsCard.tsx
@@ -1,5 +1,5 @@
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Task } from "@/types/task";
+import { Task } from "@/features/tasks/types";
 
 export const TaskApprovalsCard = ({ task }: { task: Task }) => {
   return (
