+++ frontend/src/features/tasks/components/TaskCommentsCard.tsx
@@ -1,5 +1,5 @@
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Task } from "@/types/task";
+import { Task } from "@/features/tasks/types";
 
 export const TaskCommentsCard = ({ task }: { task: Task }) => {
   return (
