+++ frontend/src/features/tasks/components/TaskModal.tsx
@@ -13,7 +13,7 @@ import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import { taskListSchema } from "@/features/tasks/datatable/TaskListSchema";
 import { useTaskMutations } from "@/features/tasks/hooks/useTaskMutation";
-import { Task, TaskCategory, TaskFrequency } from "@/types/task";
+import { Task, TaskCategory, TaskFrequency } from "@/features/tasks/types";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useEffect, useMemo } from "react";
 import { useForm } from "react-hook-form";
