+++ frontend/src/features/tasks/datatable/TaskListColumns.tsx
@@ -10,7 +10,7 @@ import { TaskFrequencyCell } from "@/features/tasks/datatable/TaskFrequencyCell"
 import { TaskRowActions } from "@/features/tasks/datatable/TaskRowActions";
 
 import { enumToArray } from "@/common/utils/enumToArray";
-import { Task, TaskCategory, TaskFrequency } from "@/types/task";
+import { Task, TaskCategory, TaskFrequency } from "@/features/tasks/types";
 import { ColumnDef } from "@tanstack/react-table";
 
 export const taskListColumns: ColumnDef<Task>[] = [
