--- frontend/src/types/navigation.ts
@@ -1,6 +1,0 @@
-export interface NavigationItem {
-  title: string;
-  label: string;
-  href: string;
-  icon?: React.ComponentType;
-}
